Added some Unit tests

+++ WorkoutTrackerAPI/Repositories/BaseRepository.cs
@@ -22,6 +22,9 @@ public class BaseRepository<T> : IDisposable, IBaseRepository<T>
 
     public async Task<T> AddAsync(T model)
     {
+        if (model.Id != 0)
+            throw new DbUpdateException($"Entity of type {typeof(T).Name} should not have an ID assigned.");
+
         T? existingModel = await GetByIdAsync(model.Id);
 
         if (existingModel is null)
@@ -36,25 +39,32 @@ public class BaseRepository<T> : IDisposable, IBaseRepository<T>
 
     public async Task AddRangeAsync(IEnumerable<T> entities)
     {
+        if (entities.Any(e => e.Id != 0))
+            throw new DbUpdateException($"New entities of type {typeof(T).Name} should not have an ID assigned.");
+
         await dbSet.AddRangeAsync(entities);
         await SaveChangesAsync();
     }
 
     public async Task RemoveAsync(long key)
     {
+        if (key <= 0)
+            throw new DbUpdateException($"Entity of type {typeof(T).Name} must have a positive ID to be removed.");
+
         T? model = await GetByIdAsync(key);
 
-        if (model is not null)
-        {
-            dbSet.Remove(model);
-            await SaveChangesAsync();
-        }
+        if (model is null)
+            throw new NotFoundException(typeof(T).Name);
 
-        throw new NotFoundException(typeof(T).Name);
+        dbSet.Remove(model);
+        await SaveChangesAsync();
     }
 
     public async Task RemoveRangeAsync(IEnumerable<T> entities)
     {
+        if (entities.Any(e => e.Id <= 0))
+            throw new DbUpdateException($"Entities of type {typeof(T).Name} must have a positive ID to be removed.");
+
         dbSet.RemoveRange(entities);
         await SaveChangesAsync();
     }
@@ -73,6 +83,9 @@ public class BaseRepository<T> : IDisposable, IBaseRepository<T>
 
     public async Task UpdateAsync(T model)
     {
+        if (model.Id < 0)
+            throw new DbUpdateException($"Modified entities of type {typeof(T).Name} must have a positive ID.");
+
         dbSet.Update(model);
         await SaveChangesAsync();
     }
